ci: run the ci/ script tests, and let a pinned release branch keep its notes
Three things, all found by asking how releases actually ship rather than how
the docs say they do.

1. ci/*.test.mjs ran nowhere. Only ci/validate-release-lines.test.mjs was
   wired into CI, by exact filename, inside release-lines-guard.yml. So
   ci/dist-tag-all.test.mjs -- the suite covering the script that moves npm's
   `latest` dist-tag at certification -- was authored, committed, and never
   once executed. Same shape as the PG-parity script that was correct,
   complete, and wired into nothing.

   ci-scripts.yml now globs ci/*.test.mjs, so a new test in that directory is
   picked up by existing rather than needing its own wiring.

2. To keep that job install-free, the back-merge logic moves to ci/back-merge.mjs,
   which imports only Node's stdlib. `simple-git` stays in the thin entry points
   (merge_main.mjs, merge_main_and_update_lock.mjs), which is where it was
   actually needed -- mergeMainIntoNext/refreshLockfile already took their client
   as a parameter. The whole ci/ suite therefore runs in ~10s with no
   `pnpm install`, matching how release-lines-guard.yml already worked. The test
   file is renamed to back-merge.test.mjs to track its module.

3. generate-release-notes.yml now also fires for `release/*` heads, not just
   `next`. Its gate exists to stop an ordinary PR into main having its title and
   description destructively overwritten, which is worth keeping -- but a release
   PR does not have to come from `next`. Shipping a specific validated commit
   means pushing it to a release/* branch and opening the PR from there, and
   without this clause doing so silently costs you the generated PR body. That is
   a direct incentive to merge the tip and hope, i.e. to reintroduce exactly the
   drift this branch removes. `release/*` is already the repo's release-branch
   convention (release/v5.51-prep, release/v5.50-prep-2, ...) and no ordinary PR
   into main uses it, so the overwrite risk is unchanged.

Also corrects the framing in DEPLOYMENT.md. The first pass described the
manual-PR path as the thing you fall back to when you cannot reach the button.
That is backwards for anyone with 5.x muscle memory: the Edge button is new in
the 6.x era and has never been dispatched, while every 5.x release shipped
through a next -> main PR, each with a release/*-prep branch feeding it. Both
paths are first-class and the doc now says so.

node --test "ci/*.test.mjs": 46 passing (10 back-merge, 36 pre-existing that had
never run in CI). Both workflow files parse.
